add Service and Change some attributes

+++ src/main/java/com/example/demo/model/Category.java
@@ -7,11 +7,11 @@ import javax.persistence.Id;
 @Entity
 public class Category {
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -26,13 +26,13 @@ public class Category {
     public Category() {
     }
 
-    public Category(int id, String name) {
+    public Category(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
     @Id
-    int id;
+    Long id;
 
     @Column(name = "name")
     String name;
